Hangman: Support optional timers
https://www.smogon.com/forums/threads/add-a-timer-to-hangmans.3765293/ 

Self explanatory suggestion, added a timer that ends the game of hangman when completed. I think the code is pretty clean besides the fix in addrandom. since the constructor follows the format of timer after the hint, random forces a timer integer which is fine since I can just use 0 for no timer by default but then the parse for the tag (hint) strings later in addrandom, flags the second tag as a timer when it's not supposed to be. I tried a lot of ideas and the only one that worked well was separating the hints (args) from being validated in the hangman class. lmk if there's a better way than what I did and I'll commit suggestion, admittedly that part was the most difficult for me but in my testing there's no issues. 

The 60 second timer was suggested by Dawoblefet but I'm open to making the max timer longer.