Hangman: Support optional timer
https://www.smogon.com/forums/threads/add-a-timer-to-hangmans.3765293/ 

Self explanatory, added a timer that ends the game of hangman when completed. I think the code is pretty clean besides the fix in addrandom. since the constructor follows the format of timer after the hint, random forces a timer integer which is fine since I can just use 0 for no timer by default but then the parse for the tag (hint) strings later in addrandom flags any second tag as a timer. I tried some ideas and the only one that worked well was allowing only the phrase specifically to be validated with the hangman class, and instead having a separate validate for hints. I think that can definitely be improved though.